First step: headers
The scribes are now correctly gathered for the header part; `scribe-` statements are ignored.

There was also a subtle bugs with `regret` vs. `regrets`: these were handled on subsequent runs on the log lines, which made it ineffective in some cases. The variants are now handled in the same run...