Minor changes
Main item: because the action handling is done via post-process, and that one has a built-in check to perform actions on a specific minutes only once, it is not necessary to check whether the action's issue has already been recorded. This means the storage of the separate action id has become unnecessary.